Revert "Refactor/search agents organize"
