--- akd/agents/litsearch/components/__init__.py
@@ -1,18 +1,0 @@
-"""
-Embedded Components for Literature Search Agents
-
-This module contains internal components that are embedded within literature search agents,
-providing deep integration of research workflow capabilities.
-"""
-
-from .triage import TriageComponent
-from .clarification import ClarificationComponent  
-from .instruction_builder import InstructionBuilderComponent
-from .research_synthesis import ResearchSynthesisComponent
-
-__all__ = [
-    "TriageComponent",
-    "ClarificationComponent",
-    "InstructionBuilderComponent", 
-    "ResearchSynthesisComponent",
-]
